[backend,shared] Move Either and Pair to shared utils
