+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/access/BenchmarkReadAccess.java
@@ -28,7 +28,7 @@ import de.aaaaaaah.velcom.backend.access.entities.sources.TarSource;
 import de.aaaaaaah.velcom.backend.access.exceptions.NoSuchRunException;
 import de.aaaaaaah.velcom.backend.storage.db.DBReadAccess;
 import de.aaaaaaah.velcom.backend.storage.db.DatabaseStorage;
-import de.aaaaaaah.velcom.backend.util.Either;
+import de.aaaaaaah.velcom.shared.util.Either;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
