+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/access/entities/Measurement.java
@@ -1,6 +1,6 @@
 package de.aaaaaaah.velcom.backend.access.entities;
 
-import de.aaaaaaah.velcom.backend.util.Either;
+import de.aaaaaaah.velcom.shared.util.Either;
 import java.util.Objects;
 
 /**
