+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/access/entities/Run.java
@@ -2,7 +2,7 @@ package de.aaaaaaah.velcom.backend.access.entities;
 
 import de.aaaaaaah.velcom.backend.access.entities.sources.CommitSource;
 import de.aaaaaaah.velcom.backend.access.entities.sources.TarSource;
-import de.aaaaaaah.velcom.backend.util.Either;
+import de.aaaaaaah.velcom.shared.util.Either;
 import java.time.Instant;
 import java.util.Collection;
 import java.util.Objects;
