+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/access/entities/benchmark/NewMeasurement.java
@@ -7,7 +7,7 @@ import de.aaaaaaah.velcom.backend.access.entities.MeasurementError;
 import de.aaaaaaah.velcom.backend.access.entities.MeasurementValues;
 import de.aaaaaaah.velcom.backend.access.entities.RunId;
 import de.aaaaaaah.velcom.backend.access.entities.Unit;
-import de.aaaaaaah.velcom.backend.util.Either;
+import de.aaaaaaah.velcom.shared.util.Either;
 import java.util.Optional;
 import javax.annotation.Nullable;
