+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/data/linearlog/CommitAccessBasedLinearLog.java
@@ -7,7 +7,7 @@ import de.aaaaaaah.velcom.backend.access.entities.BranchName;
 import de.aaaaaaah.velcom.backend.access.entities.Commit;
 import de.aaaaaaah.velcom.backend.access.entities.RepoId;
 import de.aaaaaaah.velcom.backend.access.exceptions.CommitLogException;
-import de.aaaaaaah.velcom.backend.util.Pair;
+import de.aaaaaaah.velcom.shared.util.Pair;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.Optional;
