+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/restapi/endpoints/utils/EndpointUtils.java
@@ -17,8 +17,8 @@ import de.aaaaaaah.velcom.backend.restapi.jsonobjects.JsonCommitDescription;
 import de.aaaaaaah.velcom.backend.restapi.jsonobjects.JsonResult;
 import de.aaaaaaah.velcom.backend.restapi.jsonobjects.JsonRun;
 import de.aaaaaaah.velcom.backend.restapi.jsonobjects.JsonSource;
-import de.aaaaaaah.velcom.backend.util.Either;
-import de.aaaaaaah.velcom.backend.util.Pair;
+import de.aaaaaaah.velcom.shared.util.Either;
+import de.aaaaaaah.velcom.shared.util.Pair;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Arrays;
