+++ backend/backend/src/test/java/de/aaaaaaah/velcom/backend/util/EitherTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import de.aaaaaaah.velcom.shared.util.Either;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
