+++ backend/backend/src/test/java/de/aaaaaaah/velcom/backend/util/PairTest.java
@@ -3,6 +3,7 @@ package de.aaaaaaah.velcom.backend.util;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
+import de.aaaaaaah.velcom.shared.util.Pair;
 import org.junit.jupiter.api.Test;
 
 class PairTest {
